+++ test/eca/features/prompt_test.clj
@@ -7,7 +7,7 @@
 (deftest build-instructions-test
   (testing "Should create instructions with rules, contexts, and behavior"
     (let [refined-contexts [{:type :file :path "foo.clj" :content "(ns foo)"}
-                            {:type :file :path "bar.clj" :content "(def a 1)" :partial true}
+                            {:type :file :path "bar.clj" :content "(def a 1)" :lines-range {:start 1 :end 1}}
                             {:type :repoMap}
                             {:type :mcpResource :uri "custom://my-resource" :content "some-cool-content"}]
           rules [{:name "rule1" :content "First rule"}
@@ -19,16 +19,16 @@
       (is (string/includes? result "<rules description=\"Rules defined by user\">"))
       (is (string/includes? result "<rule name=\"rule1\">First rule</rule>"))
       (is (string/includes? result "<rule name=\"rule2\">Second rule</rule>"))
-      (is (string/includes? result "<contexts description=\"Manually Provided by user. Their content is current and accurate. You MUST use this information first before using tools to read them.\">"))
+      (is (string/includes? result "<contexts description=\"User-Provided Snippet. This content is current and accurate. Treat this as sufficient context for answering the query.\">"))
       (is (string/includes? result "<file path=\"foo.clj\">(ns foo)</file>"))
-      (is (string/includes? result "<file partial=true path=\"bar.clj\">...\n(def a 1)\n...</file>"))
+      (is (string/includes? result "<file line-start=1 line-end=1 path=\"bar.clj\">(def a 1)</file>"))
       (is (string/includes? result "<repoMap description=\"Workspaces structure in a tree view, spaces represent file hierarchy\" >TREE</repoMap>"))
       (is (string/includes? result "<resource uri=\"custom://my-resource\">some-cool-content</resource>"))
       (is (string/includes? result "</contexts>"))
       (is (string? result))))
   (testing "Should create instructions with rules, contexts, and plan behavior"
     (let [refined-contexts [{:type :file :path "foo.clj" :content "(ns foo)"}
-                            {:type :file :path "bar.clj" :content "(def a 1)" :partial true}
+                            {:type :file :path "bar.clj" :content "(def a 1)" :lines-range {:start 1 :end 1}}
                             {:type :repoMap}
                             {:type :mcpResource :uri "custom://my-resource" :content "some-cool-content"}]
           rules [{:name "rule1" :content "First rule"}
@@ -40,9 +40,9 @@
       (is (string/includes? result "<rules description=\"Rules defined by user\">"))
       (is (string/includes? result "<rule name=\"rule1\">First rule</rule>"))
       (is (string/includes? result "<rule name=\"rule2\">Second rule</rule>"))
-      (is (string/includes? result "<contexts description=\"Manually Provided by user. Their content is current and accurate. You MUST use this information first before using tools to read them.\">"))
+      (is (string/includes? result "<contexts description=\"User-Provided Snippet. This content is current and accurate. Treat this as sufficient context for answering the query.\">"))
       (is (string/includes? result "<file path=\"foo.clj\">(ns foo)</file>"))
-      (is (string/includes? result "<file partial=true path=\"bar.clj\">...\n(def a 1)\n...</file>"))
+      (is (string/includes? result "<file line-start=1 line-end=1 path=\"bar.clj\">(def a 1)</file>"))
       (is (string/includes? result "<repoMap description=\"Workspaces structure in a tree view, spaces represent file hierarchy\" >TREE</repoMap>"))
       (is (string/includes? result "<resource uri=\"custom://my-resource\">some-cool-content</resource>"))
       (is (string/includes? result "</contexts>"))